+++ finagle-core/src/test/scala/com/twitter/finagle/ssl/server/SslServerEngineFactoryTest.scala
@@ -2,11 +2,8 @@ package com.twitter.finagle.ssl.server
 
 import com.twitter.finagle.ssl.{CipherSuites, ClientAuth, Engine, Protocols}
 import javax.net.ssl.{SSLContext, SSLEngine}
-import org.junit.runner.RunWith
 import org.scalatest.FunSuite
-import org.scalatest.junit.JUnitRunner
 
-@RunWith(classOf[JUnitRunner])
 class SslServerEngineFactoryTest extends FunSuite {
 
   private[this] def createTestSslContext(): SSLContext = {